now, it can run in one mobile phone

+++ src/dse/DSEInterface.java
@@ -2,7 +2,7 @@ package dse;
 
 public interface DSEInterface {
 	
-	public void setAccelerate(double x,double y,double z);
+	public void setAccelerate(float x,float y,float z);
 	public void updateDSEState(String data);
 	public void dataProcessFromGame(int type,String data);
 }
